Added inferred zone closure/inactivity for Honeywell panels
Honeywell panels don't update zone closures/inactivity properly/timely.  I added code to PartitionReady() that infers a closed/inactive status on contacts and motions when the partition is ready.  It's not perfect as a bypassed open/active zone would seemingly get updated to closed/inactive in HE when the partition is ready (not sure how to get around this at the moment).