Honeywell fix for Zone State Change Bug #14
This fixes a bug with Zone State Change for Honeywell.  Previously, if you had a large # of zones (not sure how many it takes to break), it will log errors when the zone state changes for those higher zones.  The code has been refactored to eliminate the Integer.parseInt and instead use a byte array which would not have the same limitations as an integer.